[refactor] Refactored crud
